Don't include meeting.json in zip of originals.
This JSON file is only used for Grimlock import, but as the import for
originals is no longer possible it does not make sense to include the
metadata in the zip.
