+++ changes/CA-1732.other
@@ -1,0 +1,1 @@
+No longer include `meetings.json` metadata file in ZIP download of original files. [deiferni]
